+++ src/Psd/PsdDecompressRle.h
@@ -10,7 +10,7 @@ namespace imageUtil
 {
 	/// \ingroup ImageUtil
 	/// Decompresses a block of RLE encoded data using the PackBits (http://en.wikipedia.org/wiki/PackBits) algorithm.
-	void DecompressRle(const uint8_t* PSD_RESTRICT src, unsigned int srcSize, uint8_t* PSD_RESTRICT dest, unsigned int size);
+	void DecompressRle(const uint8_t* PSD_RESTRICT src, uint64_t srcSize, uint8_t* PSD_RESTRICT dest, unsigned int size);
 
 	/// \ingroup ImageUtil
 	/// Compresses a block of data to RLE encoded data using the PackBits (http://en.wikipedia.org/wiki/PackBits) algorithm.
